+++ main_cpu.cpp
@@ -5,10 +5,10 @@
 #include "layers/dense.h"
 #include "layers/pooling.h"
 
-void test_dense(double *input) {
+void test_dense(float *input) {
     Dense d(6, 10);
     d.dump();
-    double *output = d.forward(input);
+    float *output = d.forward(input);
 
     printf("Output: ");
     for (int i=0; i<10; i++) printf("%9.6lf ", output[i]);
@@ -21,10 +21,10 @@ void test_dense(double *input) {
     d.dump();
 }
 
-void test_pooling(double *input, int c, int h, int w, int s) {
+void test_pooling(float *input, int c, int h, int w, int s) {
     Pooling p(c, h, w, s);
     p.dump();
-    double *output = p.forward(input);
+    float *output = p.forward(input);
 
     printf("Output:\n");
     for (int i=0; i<c; i++) {
@@ -45,7 +45,7 @@ void test_pooling(double *input, int c, int h, int w, int s) {
         printf("\n");
     }
 
-    double *d = p.backward(output, 0.1);
+    float *d = p.backward(output, 0.1);
     printf("Output loss:\n");
     for (int i=0; i<c; i++) {
         for (int j=0; j<h; j++) {
@@ -56,11 +56,11 @@ void test_pooling(double *input, int c, int h, int w, int s) {
     }
 }
 
-void test_conv(double *input, int c, int h, int w, int oc, int k, int s, int p) {
+void test_conv(float *input, int c, int h, int w, int oc, int k, int s, int p) {
     Conv conv(c, h, w, oc, k, s, p);
     conv.dump();
 
-    double *output = conv.forward(input);
+    float *output = conv.forward(input);
     int oh = (h+2*p-k)/s+1;
     int ow = (w+2*p-k)/s+1;
     printf("Output:\n");
@@ -73,7 +73,7 @@ void test_conv(double *input, int c, int h, int w, int oc, int k, int s, int p) 
     }
 
     for (int i=0; i<oc*oh*ow; i++) output[i] = 2;
-    double *d = conv.backward(output, 0.1);
+    float *d = conv.backward(output, 0.1);
     conv.dump();
     printf("Output loss:\n");
     for (int i=0; i<c; i++) {
@@ -88,7 +88,7 @@ void test_conv(double *input, int c, int h, int w, int oc, int k, int s, int p) 
 int main() {
     rand_init();
     int c = 1, w = 5, h = 5;
-    double *input = (double*)malloc(sizeof(double)*(c*h*w));
+    float *input = (float*)malloc(sizeof(float)*(c*h*w));
     for (int i=0; i<c*h*w; i++) input[i] = 1;//randn();
     printf("Input:\n");
     for (int i=0; i<c; i++) {
